backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pymongo==4.6.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyMuPDF==1.23.8
llama-cpp-python==0.2.20
numpy==1.24.3
pydantic==2.5.0
python-dotenv==1.0.0
motor==3.3.2
aiofiles==23.2.1
